refactor(runtime-http): Rehome HTTP seam types
Move the remaining runtime-owned connectivity and password-form seams out of the endpoint-owned package tree and into runtime.http packages.

Update runtime call sites, focused tests, and package documentation while preserving the legacy connectivity override property key and delegated password-form rendering behavior.
