+++ src/main/java/network/crypta/node/subsystem/NodeStorageSubsystem.java
@@ -47,8 +47,8 @@ import network.crypta.node.stats.NotAvailNodeStoreStats;
 import network.crypta.node.stats.StoreCallbackStats;
 import network.crypta.runtime.alerts.UserAlert;
 import network.crypta.runtime.alerts.UserAlertManagerStoreAlertSink;
-import network.crypta.runtime.endpoints.http.security.PasswordFormPageRenderer;
-import network.crypta.runtime.endpoints.http.security.PasswordPromptOptions;
+import network.crypta.runtime.http.security.PasswordFormPageRenderer;
+import network.crypta.runtime.http.security.PasswordPromptOptions;
 import network.crypta.store.BlockMetadata;
 import network.crypta.store.CHKStore;
 import network.crypta.store.FreenetStore;
