+++ src/main/java/network/crypta/runtime/http/security/PasswordFormPageRenderer.java
@@ -1,4 +1,4 @@
-package network.crypta.runtime.endpoints.http.security;
+package network.crypta.runtime.http.security;
 
 import java.util.Objects;
 import network.crypta.clients.http.PasswordFormOptions;
